backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
python-decouple>=3.8
psycopg2-binary>=2.9.0
django-cors-headers>=4.3.0
Pillow>=10.0.0